Move all the practice components to a separate directory.
